requirements.txt
gevent==24.2.1
psycogreen==1.0.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
redis==5.0.4
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4
python-dotenv==1.0.1
pytest==8.2.1